requirements.txt
python-dotenv==1.0.1
python-telegram-bot==21.10
selenium==4.29.0

